requirements.txt
streamlit>=1.33.0
requests>=2.31.0
python-dotenv>=1.0.1
openai>=1.30.0
numpy>=1.26.0
scikit-learn>=1.4.0
